+++ jsoneditor/js/util.js
@@ -1,8 +1,6 @@
 // create namespace
 util = {};
 
-// Internet Explorer 8 and older does not support Array.indexOf,
-// so we define it here in that case
 // http://soledadpenades.com/2007/05/17/arrayindexof-in-internet-explorer/
 if(!Array.prototype.indexOf) {
     Array.prototype.indexOf = function(obj){
@@ -15,8 +13,6 @@ if(!Array.prototype.indexOf) {
     }
 }
 
-// Internet Explorer 8 and older does not support Array.forEach,
-// so we define it here in that case
 // https://developer.mozilla.org/en-US/docs/JavaScript/Reference/Global_Objects/Array/forEach
 if (!Array.prototype.forEach) {
     Array.prototype.forEach = function(fn, scope) {
@@ -26,12 +22,19 @@ if (!Array.prototype.forEach) {
     }
 }
 
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/isArray
+if(!Array.isArray) {
+    Array.isArray = function (vArg) {
+        return Object.prototype.toString.call(vArg) === "[object Array]";
+    };
+}
+
 /**
  * Parse JSON using the parser built-in in the browser.
  * On exception, the jsonString is validated and a detailed error is thrown.
  * @param {String} jsonString
  */
-util.parse = function (jsonString) {
+util.parse = function parse(jsonString) {
     try {
         return JSON.parse(jsonString);
     }
@@ -49,7 +52,7 @@ util.parse = function (jsonString) {
  * @param {String} jsonString   String with an (invalid) JSON object
  * @throws Error
  */
-util.validate = function (jsonString) {
+util.validate = function validate(jsonString) {
     if (typeof(jsonlint) != 'undefined') {
         jsonlint.parse(jsonString);
     }
@@ -64,7 +67,7 @@ util.validate = function (jsonString) {
  * @param {Object} b
  * @return {Object} a
  */
-util.extend = function (a, b) {
+util.extend = function extend(a, b) {
     for (var prop in b) {
         if (b.hasOwnProperty(prop)) {
             a[prop] = b[prop];
@@ -78,7 +81,7 @@ util.extend = function (a, b) {
  * @param {Object} a
  * @return {Object} a
  */
-util.clear = function (a) {
+util.clear = function clear (a) {
     for (var prop in a) {
         if (a.hasOwnProperty(prop)) {
             delete a[prop];
@@ -91,10 +94,41 @@ util.clear = function (a) {
  * Output text to the console, if console is available
  * @param {...*} args
  */
-util.log = function(args) {
+util.log = function log (args) {
     if (console && typeof console.log === 'function') {
         console.log.apply(console, arguments);
     }
+};
+
+/**
+ * Get the type of an object
+ * @param {*} object
+ * @return {String} type
+ */
+util.type = function type (object) {
+    if (object === null) {
+        return 'null';
+    }
+    if (object === undefined) {
+        return 'undefined';
+    }
+    if ((object instanceof Number) || (typeof object === 'number')) {
+        return 'number';
+    }
+    if ((object instanceof String) || (typeof object === 'string')) {
+        return 'string';
+    }
+    if ((object instanceof Boolean) || (typeof object === 'boolean')) {
+        return 'boolean';
+    }
+    if ((object instanceof RegExp) || (typeof object === 'regexp')) {
+        return 'regexp';
+    }
+    if (Array.isArray(object)) {
+        return 'array';
+    }
+
+    return 'object';
 };
 
 /**
@@ -103,7 +137,7 @@ util.log = function(args) {
  * @param {String} text
  */
 var isUrlRegex = /^https?:\/\/\S+$/;
-util.isUrl = function (text) {
+util.isUrl = function isUrl (text) {
     return (typeof text == 'string' || text instanceof String) &&
         isUrlRegex.test(text);
 };
@@ -114,7 +148,7 @@ util.isUrl = function (text) {
  * @return {Number} left    The absolute left position of this element
  *                          in the browser page.
  */
-util.getAbsoluteLeft = function (elem) {
+util.getAbsoluteLeft = function getAbsoluteLeft(elem) {
     var left = elem.offsetLeft;
     var body = document.body;
     var e = elem.offsetParent;
@@ -132,7 +166,7 @@ util.getAbsoluteLeft = function (elem) {
  * @return {Number} top    The absolute top position of this element
  *                          in the browser page.
  */
-util.getAbsoluteTop = function (elem) {
+util.getAbsoluteTop = function getAbsoluteTop(elem) {
     var top = elem.offsetTop;
     var body = document.body;
     var e = elem.offsetParent;
@@ -149,7 +183,7 @@ util.getAbsoluteTop = function (elem) {
  * @param {Event} event
  * @return {Number} mouseY
  */
-util.getMouseY = function (event) {
+util.getMouseY = function getMouseY(event) {
     var mouseY;
     if ('pageY' in event) {
         mouseY = event.pageY;
@@ -167,7 +201,7 @@ util.getMouseY = function (event) {
  * @param {Event} event
  * @return {Number} mouseX
  */
-util.getMouseX = function (event) {
+util.getMouseX = function getMouseX(event) {
     var mouseX;
     if ('pageX' in event) {
         mouseX = event.pageX;
@@ -184,7 +218,7 @@ util.getMouseX = function (event) {
  * Get the window height
  * @return {Number} windowHeight
  */
-util.getWindowHeight = function () {
+util.getWindowHeight = function getWindowHeight() {
     if ('innerHeight' in window) {
         return window.innerHeight;
     }
@@ -200,7 +234,7 @@ util.getWindowHeight = function () {
  * @param {Element} elem
  * @param {String} className
  */
-util.addClassName = function(elem, className) {
+util.addClassName = function addClassName(elem, className) {
     var classes = elem.className.split(' ');
     if (classes.indexOf(className) == -1) {
         classes.push(className); // add the class to the array
@@ -213,7 +247,7 @@ util.addClassName = function(elem, className) {
  * @param {Element} elem
  * @param {String} className
  */
-util.removeClassName = function(elem, className) {
+util.removeClassName = function removeClassName(elem, className) {
     var classes = elem.className.split(' ');
     var index = classes.indexOf(className);
     if (index != -1) {
@@ -227,7 +261,7 @@ util.removeClassName = function(elem, className) {
  * the formatting from the div itself is not stripped, only from its childs.
  * @param {Element} divElement
  */
-util.stripFormatting = function (divElement) {
+util.stripFormatting = function stripFormatting(divElement) {
     var childs = divElement.childNodes;
     for (var i = 0, iMax = childs.length; i < iMax; i++) {
         var child = childs[i];
@@ -261,7 +295,7 @@ util.stripFormatting = function (divElement) {
  * http://stackoverflow.com/questions/1125292/how-to-move-cursor-to-end-of-contenteditable-entity
  * @param {Element} contentEditableElement   A content editable div
  */
-util.setEndOfContentEditable = function (contentEditableElement) {
+util.setEndOfContentEditable = function setEndOfContentEditable(contentEditableElement) {
     var range, selection;
     if(document.createRange) {//Firefox, Chrome, Opera, Safari, IE 9+
         range = document.createRange();//Create a range (a range is a like the selection but invisible)
@@ -284,7 +318,7 @@ util.setEndOfContentEditable = function (contentEditableElement) {
  * http://stackoverflow.com/a/3806004/1262753
  * @param {Element} contentEditableElement   A content editable div
  */
-util.selectContentEditable = function (contentEditableElement) {
+util.selectContentEditable = function selectContentEditable(contentEditableElement) {
     if (!contentEditableElement || contentEditableElement.nodeName != 'DIV') {
         return;
     }
@@ -308,7 +342,7 @@ util.selectContentEditable = function (contentEditableElement) {
  * http://stackoverflow.com/questions/4687808/contenteditable-selected-text-save-and-restore
  * @return {Range | TextRange | null} range
  */
-util.getSelection = function () {
+util.getSelection = function getSelection() {
     if (window.getSelection) {
         var sel = window.getSelection();
         if (sel.getRangeAt && sel.rangeCount) {
@@ -325,7 +359,7 @@ util.getSelection = function () {
  * http://stackoverflow.com/questions/4687808/contenteditable-selected-text-save-and-restore
  * @param {Range | TextRange | null} range
  */
-util.setSelection = function (range) {
+util.setSelection = function setSelection(range) {
     if (range) {
         if (window.getSelection) {
             var sel = window.getSelection();
@@ -346,7 +380,7 @@ util.setSelection = function (range) {
  *                                                   selected text element
  *                          Returns null if no text selection is found
  */
-util.getSelectionOffset = function () {
+util.getSelectionOffset = function getSelectionOffset() {
     var range = util.getSelection();
 
     if (range && 'startOffset' in range && 'endOffset' in range &&
@@ -371,7 +405,7 @@ util.getSelectionOffset = function () {
  *                              {Number} startOffset
  *                              {Number} endOffset
  */
-util.setSelectionOffset = function (params) {
+util.setSelectionOffset = function setSelectionOffset(params) {
     if (document.createRange && window.getSelection) {
         var selection = window.getSelection();
         if(selection) {
@@ -395,7 +429,7 @@ util.setSelectionOffset = function (params) {
  * @param {Object} [buffer]
  * @return {String} innerText
  */
-util.getInnerText = function (element, buffer) {
+util.getInnerText = function getInnerText(element, buffer) {
     var first = (buffer == undefined);
     if (first) {
         buffer = {
@@ -466,7 +500,7 @@ util.getInnerText = function (element, buffer) {
  * Source: http://msdn.microsoft.com/en-us/library/ms537509(v=vs.85).aspx
  * @return {Number} Internet Explorer version, or -1 in case of an other browser
  */
-util.getInternetExplorerVersion = function() {
+util.getInternetExplorerVersion = function getInternetExplorerVersion() {
     if (_ieVersion == -1) {
         var rv = -1; // Return value assumes failure.
         if (navigator.appName == 'Microsoft Internet Explorer')
@@ -500,7 +534,7 @@ var _ieVersion = -1;
  * @param {boolean}     [useCapture] false by default
  * @return {function}   the created event listener
  */
-util.addEventListener = function (element, action, listener, useCapture) {
+util.addEventListener = function addEventListener(element, action, listener, useCapture) {
     if (element.addEventListener) {
         if (useCapture === undefined)
             useCapture = false;
@@ -528,7 +562,7 @@ util.addEventListener = function (element, action, listener, useCapture) {
  * @param {function} listener  The listener function
  * @param {boolean}  [useCapture]   false by default
  */
-util.removeEventListener = function(element, action, listener, useCapture) {
+util.removeEventListener = function removeEventListener(element, action, listener, useCapture) {
     if (element.removeEventListener) {
         // non-IE browsers
         if (useCapture === undefined)
@@ -550,7 +584,7 @@ util.removeEventListener = function(element, action, listener, useCapture) {
  * Stop event propagation
  * @param {Event} event
  */
-util.stopPropagation = function (event) {
+util.stopPropagation = function stopPropagation(event) {
     if (!event) {
         event = window.event;
     }
@@ -568,7 +602,7 @@ util.stopPropagation = function (event) {
  * Cancels the event if it is cancelable, without stopping further propagation of the event.
  * @param {Event} event
  */
-util.preventDefault = function (event) {
+util.preventDefault = function preventDefault(event) {
     if (!event) {
         event = window.event;
     }